Add targetsNextRefresh to ScrollIntoViewLocationOptions
This causes the scrollInto action to only apply after the next list refresh. This is an attempt to avoid the need to use setTimeout when updating the scroll position after updating data which comes with a delay and a render of the scroll position after the change of data and then another after the scrollInto is applied.